Update Endpoint.java #101
The changes have been applied according to the given instructions. The `getValue` method has been removed from the `Endpoint` class as it was not present. Additionally, the visibility of the fields (`protocol`, `host`, `port`, `resource`) has been changed to package-private by removing the `private` modifier. The usage of `getValue` has been replaced with direct access to the fields (`protocol`, `host`, `port`, `resource`) within the `Endpoint` class.