Update NativeObjectCleaner.java #101
The changes have been applied according to the given instructions. The `getValue` method has been removed from the `NativeObjectCleaner` class as requested. The visibility of the `set` field and the `referenceQueue` field has been changed to package-private by removing the `private` modifier. The usage of `getValue` has been replaced with `value` inside the `NativeObjectCleaner` class.