Update RocContext.java #101
The changes have been applied according to the given instructions. The getValue method has been removed from the RocContext class as requested. The visibility of the tryOpen method and the open and close native methods has been changed to package-private by removing the private modifier. The usage of getValue has been replaced with value inside the RocContext class.