Update RocContextConfig.java #101
The changes have been applied according to the given instructions. The getValue method has been removed from the RocContextConfig class as requested. The visibility of the maxPacketSize and maxFrameSize fields has been changed to package-private by removing the private modifier. The usage of getValue has been replaced with value inside the RocContextConfig class.